+++ __fixtures__/rules/src/modules/demo/routes.tsx
@@ -1,0 +1,1 @@
+export const demoRoutes = [{ path: '/demo', lazy: () => import('./views/DeepPage') }]