+++ __fixtures__/rules/src/modules/demo/views/WidgetPage.tsx
@@ -1,0 +1,7 @@
+export default function WidgetPage() {
+  const a = 1
+  const b = 2
+  const c = 3
+  const d = 4
+  return <div>{a + b + c + d}</div>
+}