+++ __fixtures__/rules/src/modules/demo/views/nested/DeepPage.tsx
@@ -1,0 +1,3 @@
+export default function DeepPage() {
+  return <div>deep</div>
+}